refactor(finding): extract API layer into dojo/finding/api/ [finding Phase 6,7,8,9]
