+++ dojo/finding/api/__init__.py
@@ -1,0 +1,1 @@
+path = "findings"  # noqa: RUF067
