+++ unittests/test_rest_framework.py
@@ -44,8 +44,6 @@ from dojo.api_v2.views import (
     DevelopmentEnvironmentViewSet,
     EndpointStatusViewSet,
     EndPointViewSet,
-    FindingTemplatesViewSet,
-    FindingViewSet,
     ImportLanguagesView,
     ImportScanView,
     JiraInstanceViewSet,
@@ -69,6 +67,7 @@ from dojo.asset.api.views import (
 )
 from dojo.authorization.roles_permissions import Permissions, permission_to_action
 from dojo.engagement.api.views import EngagementViewSet
+from dojo.finding.api.views import FindingTemplatesViewSet, FindingViewSet
 from dojo.location.api.endpoint_compat import V3EndpointCompatibleViewSet, V3EndpointStatusCompatibleViewSet
 from dojo.location.api.views import LocationFindingReferenceViewSet, LocationProductReferenceViewSet, LocationViewSet
 from dojo.location.models import Location, LocationFindingReference, LocationProductReference
